Reboot to Advance course

+++ app/resources/item.py
@@ -1,29 +1,32 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_claims,
+    get_jwt_identity,
+    jwt_optional,
+    fresh_jwt_required,
+)
 
 from models.item import ItemModel
 
 
 class Item(Resource):
     parser = reqparse.RequestParser()
-    parser.add_argument('price',
-        type=float,
-        required=True,
-        help="This field can't be left blank!"
+    parser.add_argument(
+        'price', type=float, required=True, help="This field can't be left blank!"
     )
-    parser.add_argument('store_id',
-        type=int,
-        required=True,
-        help="Every item needs a store id."
+    parser.add_argument(
+        'store_id', type=int, required=True, help="Every item needs a store id."
     )
-    
-    @jwt_required()
+
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             message = "An item with name '{}' already exists.".format(name)
@@ -41,12 +44,17 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required."}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-
-        return {'message': 'Item deleted'}
+            return {'message': 'Item deleted'}, 200
+        return {'message': 'Item not found'}, 404
 
     def put(self, name):
         data = Item.parser.parse_args()
@@ -57,13 +65,32 @@ class Item(Resource):
             item = ItemModel(name, **data)
         else:
             item.price = data['price']
+
         item.save_to_db()
 
-        return item.json()
-    
+        return item.json(), 200
+
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        """Here we get the JWT identity, and then if the user is logged in (we were able to get an identity)
+        we return the entire item list.
 
+        Otherwise we just return the item names.
 
+        This could be done with e.g. see orders that have been placed, but not see details about the orders
+        unless the user has logged in."""
+
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {"items": items}, 200
+
+        return (
+            {
+                'items': [item['name'] for item in items],
+                'messages': "More data available if you log in."
+            },
+            200,
+        )
